poleno/cron/requirements.txt
django-cron
